Update print statements and precision in coordinates calculations
Long double precision has been implemented for variable representation and print statements in the 'state_vector.c' file. This is done to enhance the accuracy of these variables. Simultaneously, print message details have been revised for better clarification, particularly in differentiating before and after states.

+++ c/include/math/linear_algebra.h
@@ -38,7 +38,7 @@ extern "C" {
  *  Member 'z' is the z component of the vector
  */
 typedef struct {
-    long double x, y, z;
+    double x, y, z;
 } Vec3;
 
 /**
